several changes: - chore: improve interfaces for price generation - feat: create gas, crude and power daily price generations (implementing 'ICreateDailyPricesGenerator') - chore: remove unused code  (timestamps and price generator, commodities map) - chore: add stack map and first stack (for GB) - chore: add 'IPricesGenerator' interface to be used as a closure for bespoke price generators
