+++ ctmds/domain/commodity_price/commodities/electricity/price_generators/basic_stack_based.py
@@ -6,14 +6,15 @@ from ctmds.domain.commodity_price.commodities.electricity.generation_stack.deman
 from ctmds.domain.commodity_price.commodities.electricity.generation_stack.interfaces import (
     IDemandProfileGenerator,
 )
-from ctmds.domain.commodity_price.commodities.electricity.generation_stack.stack import (
-    GenerationStack,
+from ctmds.domain.commodity_price.commodities.electricity.price_generators.stacks import (
+    StackMap,
 )
+from ctmds.domain.constants import CountryCodes
+from ctmds.domain.data_generators.daily_timestamps import get_day_periods
 
 
 def stack_generation_based_price_generator(
-    periods: int,
-    stack: GenerationStack,
+    country_code: CountryCodes,
     date: datetime,
     demand_profile: IDemandProfileGenerator = generate_demand_profile,
     seed: int | None = None,
@@ -22,8 +23,7 @@ def stack_generation_based_price_generator(
     Generate power prices using a merit-order based generation stack.
 
     Args:
-        periods: Number of periods to generate prices for
-        stack: Generation stack to use for price calculation
+        country_code: Country code to use for price calculation
         demand_profile: Function to generate demand profile
         date: Date to generate prices for
         seed: Random seed for reproducibility
@@ -33,6 +33,9 @@ def stack_generation_based_price_generator(
     """
     base_demand = 25.0
     peak_demand = 40.0
+
+    stack = StackMap.get_stack(country_code)
+    periods = get_day_periods(date, country_code)
 
     # Generate demand profile
     demands = demand_profile(
