+++ ctmds/domain/commodity_price/commodities/generators_map.py
@@ -1,19 +1,18 @@
-# from ctmds.domain.commodity_price.commodities.crude.price_generator import (
-#     oil_price_generator,
-# )
 from typing import Protocol
 
-from ctmds.domain.commodity_price.commodities.electricity.model import (
+from ctmds.domain.commodity_price.commodities.crude.price_generator import (
+    CrudePriceGenerator,
+)
+from ctmds.domain.commodity_price.commodities.electricity.price_generator import (
     ElectricityPriceGenerator,
 )
-
-# from ctmds.domain.commodity_price.commodities.natural_gas.price_generator import (
-#     gas_price_generator,
-# )
+from ctmds.domain.commodity_price.commodities.natural_gas.price_generator import (
+    GasPriceGenerator,
+)
 from ctmds.domain.commodity_price.models.price_generator.generic import (
     GenericDailyPricesGenerator,
 )
-from ctmds.domain.constants import SupportedCommodities
+from ctmds.domain.constants import CountryCodes, SupportedCommodities
 
 
 class ICreateDailyPricesGenerator(Protocol):
@@ -28,18 +27,19 @@ class GeneratorMap:
     """Map of supported commodities to their implementations."""
 
     _generators_map = {
-        # SupportedCommodities.CRUDE: oil_price_generator,
-        # SupportedCommodities.NATURAL_GAS: gas_price_generator,
-        SupportedCommodities.POWER: ElectricityPriceGenerator(),
+        SupportedCommodities.CRUDE: CrudePriceGenerator,
+        SupportedCommodities.NATURAL_GAS: GasPriceGenerator,
+        SupportedCommodities.POWER: ElectricityPriceGenerator,
     }
 
     @staticmethod
     def get_generator(
         commodity: SupportedCommodities,
+        country_code: CountryCodes,
     ) -> ICreateDailyPricesGenerator:
         """Get the commodity implementation for a given commodity."""
         generator = GeneratorMap._generators_map.get(
             commodity,
-            ElectricityPriceGenerator(),
+            ElectricityPriceGenerator,
         )
-        return generator
+        return generator(country_code)
