+++ ctmds/domain/commodity_price/models/commodity/generic.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import cast
 
 from ctmds.core.data_access.reader import GenericReader
@@ -5,7 +6,7 @@ from ctmds.core.db.session import SessionStream, get_session
 from ctmds.core.utils.filter_sort import BaseFilterParams
 from ctmds.domain import exceptions
 from ctmds.domain.commodity_price.models.commodity.interface import ICommodity
-from ctmds.domain.constants import CountryCodes
+from ctmds.domain.constants import CountryCodes, TimePeriod
 from ctmds.domain.entities import CommodityConfig
 
 
@@ -21,11 +22,14 @@ class GenericCommodity(ICommodity):
     def __init__(self):
         self.session_stream: SessionStream = get_session
 
-    def get_base_price(
+    def get_price(
         self,
         country_code: CountryCodes,
-    ) -> float:
+        period: TimePeriod,
+    ) -> Decimal:
         """Get base price for generic commodity in a country, if valid."""
+
+        # TODO: Get price for "period" from database table with daily prices
 
         valid_country_code = self.reader.get_by(
             db=next(self.session_stream()),
@@ -37,7 +41,7 @@ class GenericCommodity(ICommodity):
                 f"Country code must be one of: {', '.join(self.get_supported_countries())}"
             )
 
-        base_price = cast(float, valid_country_code.base_price)
+        base_price = cast(Decimal, valid_country_code.base_price)
         return base_price
 
     def get_supported_countries(self) -> list[CountryCodes]:
